filmpro-service/pyproject.toml
[tool.poetry]
name = "filmpro-service"
version = "0.1.0"
description = "FILMPRO Script Analysis Service"
authors = ["FILMPRO Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
pydantic = "^1.10.7"
beanie = "^1.17.0"
spacy = "^3.5.2"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.6.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
requests = "^2.28.2"
redis = "^4.5.4"
email-validator = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.2.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
asyncio_mode = "auto"